requirements.txt
Flask==3.0.0
supabase>=2.9.0
httpx>=0.27,<0.28
websockets>=15.0
python-dotenv==1.0.0
gunicorn==21.2.0
